requirements.txt
Flask
Flask-Mail
requests
setuptools
python-dotenv
requests
twine
qrcode
pillow